requirements.txt
streamlit==1.51.0
openai==1.0.0
faiss-cpu==1.12.0
pandas
numpy
scikit-learn
requests

